refactor: move BlockingQueues mem/redis into app module #1384
