+++ .gitignore
@@ -14,4 +14,5 @@ datashare-web/src/main/resources/secret.properties
 .settings
 .classpath
 datashare-cli/doc/
+datashare-tasks/doc/
 work
