+++ datashare-app/src/main/java/org/icij/datashare/mode/CommonMode.java
@@ -25,8 +25,8 @@ import org.icij.datashare.asynctasks.TaskModifier;
 import org.icij.datashare.asynctasks.TaskSupplier;
 import org.icij.datashare.asynctasks.TaskView;
 import org.icij.datashare.asynctasks.bus.amqp.AmqpInterlocutor;
-import org.icij.datashare.asynctasks.bus.memory.MemoryBlockingQueue;
-import org.icij.datashare.asynctasks.bus.redis.RedisBlockingQueue;
+import org.icij.datashare.com.queue.MemoryBlockingQueue;
+import org.icij.datashare.com.queue.RedisBlockingQueue;
 import org.icij.datashare.batch.BatchSearchRepository;
 import org.icij.datashare.cli.Mode;
 import org.icij.datashare.cli.QueueType;
